moved files and improved user functions
