+++ transcendence/api/urls.py
@@ -3,17 +3,18 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path("user", views.profilePicture, name="user"),
+    path("user", views.userController.profilePicture, name="user"),
 	
-    path("friends", views.getFriends, name="getFriends"),
-    path("getUsers/<str:search>", views.getUsers, name="getUsers"),
-    path("users/<str:userName>", views.getUser, name="getUser"),
+    path("friends", views.userController.getFriends, name="getFriends"),
+    path("searchUsers/<str:search>", views.userController.searchUsers, name="getUsers"),
+    path("users", views.userController.Users, name="getUser"),
+    path("users/<str:userName>", views.userController.getUser, name="getUser"),
 	
-    path("deleteUser", views.deleteUser, name="deleteUser"),
-    path("putUser", views.putUser, name="putUser"),
+    path("deleteUser/<str:userName>", views.userController.Users, name="deleteUser"),
+    path("updateUser/<str:userName>", views.userController.Users, name="putUser"),
 	
-    path("tournament", views.tournamentManager, name="tournament"),
-	path("tournament/<int:id>", views.tournamentManagerID, name="tournamentID"),
-    path("tournament/player", views.tournamentPlayer, name="tournament/player"),
-    path("tournament/player/<int:id>", views.tournamentManagerPlayerID, name="tournamentPlayerID"),
+    path("tournament", views.tournamentController.tournamentManager, name="tournament"),
+	path("tournament/<int:id>", views.tournamentController.tournamentManagerID, name="tournamentID"),
+    path("tournament/player", views.tournamentController.tournamentPlayer, name="tournament/player"),
+    path("tournament/player/<int:id>", views.tournamentController.tournamentManagerPlayerID, name="tournamentPlayerID"),
 ]
